kukkanettisivu/requirements.txt
asgiref==3.7.2
Babel==2.12.1
black==23.7.0
flake8==6.1.0
click==8.1.6
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg==3.1.10
python-magic==0.4.27
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
Pillow==10.1.0
pytest==7.4.3
